+++ src/components/algolia/SearchSuggestions.js
@@ -1,8 +1,8 @@
-import React, { Component } from 'react';
-import { connectMenu } from 'react-instantsearch-dom';
-import { withStyles, Chip } from '@material-ui/core'
+import React from 'react'
+import { Grid, withStyles, Chip } from '@material-ui/core';
+import { inject, observer } from "mobx-react";
 
-const style = theme => ({
+const styles = theme => ({
   suggestionsContainer: {
     textAlign: 'left',
     maxHeight: 112, // 48 * 2 + 8 + 8
@@ -33,31 +33,61 @@ const style = theme => ({
   }
 });
 
-class SearchSuggestionsComponent extends Component {
+class SearchSuggestions extends React.Component {
   constructor(props) {
     super(props);
+    this.state = {
+      facetHits: []
+    };
+
+    this.fetchSuggestions = this.fetchSuggestions.bind(this);
   }
-  
+
+  componentDidMount() {
+    this.fetchSuggestions(this.props.filters);
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if(nextProps.filters) {
+      this.fetchSuggestions(nextProps.filters);
+    }
+  }
+
+  fetchSuggestions(filters) {
+    return this.props.index.searchForFacetValues({
+      facetName: 'hashtags.tag',
+      facetQuery: '',
+      query: '',
+      facetFilters: filters.split(' AND '),
+      filters: filters || '',
+    }, (err, res) => {
+      if(!err) {
+        this.setState({facetHits: res.facetHits});
+      }
+    });
+  }
+
   shouldDisplaySuggestion(tag) {
-    return (this.props.currentFilters.search(tag) === -1);
+    return (this.props.filters.search(tag) === -1);
   }
 
   render() {
-    const { items, classes, addToFilters } = this.props;
+    const {facetHits} = this.state;
+    const {classes, addToFilters} = this.props;
 
     return (
       <div className={classes.suggestionsContainer} >
-        {items.map((item, i) => {
+        {facetHits.map((item, i) => {
           if (this.shouldDisplaySuggestion(item.value)){
             return (
               <Chip key={i} 
                     component={ (props)=>{
                       return (<div {...props}>
-                                <div className={classes.suggestionLabel}>{item.label}</div>
+                                <div className={classes.suggestionLabel}>{item.value}</div>
                                 <div className={classes.suggestionCount}>{item.count}</div>
                               </div>);
                     }}
-                    onClick={(e) => addToFilters(e, { name: item.label, tag: item.value })} 
+                    onClick={(e) => addToFilters(e, { name: item.value, tag: item.value })} 
                     className={classes.suggestion} />
             );
           }else {
@@ -69,5 +99,8 @@ class SearchSuggestionsComponent extends Component {
   }
 }
 
-const SearchSuggestions = connectMenu(SearchSuggestionsComponent);
-export default withStyles(style, { withTheme: true })(SearchSuggestions);
+export default inject('commonStore', 'organisationStore', 'authStore', 'recordStore', 'userStore')(
+  observer(
+    withStyles(styles)(SearchSuggestions)
+  )
+);
