--- src/components/algolia/SearchSuggestions2.js
@@ -1,106 +1,0 @@
-import React from 'react'
-import { Grid, withStyles, Chip } from '@material-ui/core';
-import { inject, observer } from "mobx-react";
-
-const styles = theme => ({
-  suggestionsContainer: {
-    textAlign: 'left',
-    maxHeight: 112, // 48 * 2 + 8 + 8
-    overflow: 'hidden',
-    padding: '8px 0px',
-    marginLeft: '-8px',
-  },
-  suggestion: {
-    margin: 8,
-    paddingRight: 0,
-    background: 'white',
-    color: theme.palette.secondary.dark,
-    '&:hover': {
-      background: 'rgb(220,220,220)'
-    },
-  },
-  suggestionCount: {
-    background: 'rgb(220,220,220)',
-    color: 'white',
-    borderRadius: '50%',
-    width: 32,
-    height:32,
-    textAlign: 'center',
-    lineHeight: '32px'
-  },
-  suggestionLabel: {
-    marginRight: 8,
-  }
-});
-
-class SearchSuggestions2 extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      facetHits: []
-    };
-
-    this.fetchSuggestions = this.fetchSuggestions.bind(this);
-  }
-
-  componentDidMount() {
-    this.fetchSuggestions(this.props.filters);
-  }
-
-  componentWillReceiveProps(nextProps) {
-    if(nextProps.filters) {
-      this.fetchSuggestions(nextProps.filters);
-    }
-  }
-
-  fetchSuggestions(filters) {
-    return this.props.index.searchForFacetValues({
-      facetName: 'hashtags.tag',
-      facetQuery: '',
-      query: '',
-      facetFilters: filters.split(' AND '),
-      filters: filters || '',
-    }, (err, res) => {
-      if(!err) {
-        this.setState({facetHits: res.facetHits});
-      }
-    });
-  }
-
-  shouldDisplaySuggestion(tag) {
-    return (this.props.filters.search(tag) === -1);
-  }
-
-  render() {
-    const {facetHits} = this.state;
-    const {classes, addToFilters} = this.props;
-
-    return (
-      <div className={classes.suggestionsContainer} >
-        {facetHits.map((item, i) => {
-          if (this.shouldDisplaySuggestion(item.value)){
-            return (
-              <Chip key={i} 
-                    component={ (props)=>{
-                      return (<div {...props}>
-                                <div className={classes.suggestionLabel}>{item.value}</div>
-                                <div className={classes.suggestionCount}>{item.count}</div>
-                              </div>);
-                    }}
-                    onClick={(e) => addToFilters(e, { name: item.value, tag: item.value })} 
-                    className={classes.suggestion} />
-            );
-          }else {
-            return null;
-          }
-        })}
-      </div>
-    );
-  }
-}
-
-export default inject('commonStore', 'organisationStore', 'authStore', 'recordStore', 'userStore')(
-  observer(
-    withStyles(styles)(SearchSuggestions2)
-  )
-);
